pyproject.toml
[project]
name = "fastapi-vitest"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "andrewgryan", email = "andrew.ryan@metoffice.gov.uk"},
]
dependencies = [
    "fastapi[all]>=0.111.0",
    "pytest-playwright>=0.5.0",
    "pytest>=8.2.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start.cmd = "uvicorn main:app --reload"
start.working_dir = "src/fastapi_vitest"
test.cmd = "yarn test"
test.working_dir = "tests/frontend"
